pyproject.toml
[project]
name = "ragtest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cachetools>=6.2.2",
    "chromadb>=1.3.5",
    "faiss-cpu>=1.13.1",
    "ipykernel>=7.1.0",
    "langchain>=1.1.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.1",
    "langchain-groq>=1.1.0",
    "mysql-connector-python>=9.5.0",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.26.6",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.52.1",
]
